requirements.txt
fastapi
uvicorn
anthropic
jinja2
python-dotenv
pytest
httpx
